third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/getAll-expected.txt

This is a testharness.js-based test.
PASS Calling StylePropertyMap.getAll with an unsupported property throws a TypeError
PASS Calling StylePropertyMap.getAll with a property not in the property model returns an empty list
PASS Calling StylePropertyMap.getAll with a custom property not in the property model returns an empty list
PASS Calling StylePropertyMap.getAll with a valid property returns a single element list with the correct entry
PASS StylePropertyMap.getAll is case-insensitive
FAIL Calling StylePropertyMap.getAll with a valid custom property returns a single element list with the correct entry assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Calling StylePropertyMap.getAll with a list-valued property returns all the values assert_equals: expected 2 but got 1
Harness: the test ran to completion.
